feat: add support for division and modulo
This adds support for the division and modulo operators.

The division operator is implemented as a call to the `div` function, which
is implemented in the runtime. The modulo operator is implemented as a call
to the `mod` function, which is also implemented in the runtime.

The `div` and `mod` functions are implemented in the runtime using the
`i32.div_s` and `i32.rem_s` instructions, respectively.

The `div` and `mod` functions are also added to the typechecker.